refactor: move proxmox component into components package
