+++ pulumi/components/__init__.py
@@ -1,0 +1,1 @@
+"""Reusable Pulumi component resources for CrowsNet infrastructure."""
